# openapi.yaml
openapi: 3.0.3
info:
  title: Lume Store · API Gateway
  version: 2.0.0
  description: |
    Todas as chamadas passam pelo **API Gateway**, que valida o token e repassa ao microserviço certo:
    `/api/itens` → serviço de itens, `/api/pedidos` → serviço de pedidos, `/api/usuarios` e `/api/auth` → serviço de usuários.

    1. Faça login em **POST /api/auth/login** (`cliente@lumestore.dev` / `senha123` ou `admin@lumestore.dev` / `admin123`).
    2. Copie o `token`, clique em **Authorize** e cole.
    3. Crie um pedido e veja o estoque do item diminuir.
servers:
  - url: /
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Erro:
      type: object
      properties:
        erro: { type: string }
...
    Usuario:
      type: object
      properties:
        id: { type: integer }
        nome: { type: string }
        email: { type: string }
        endereco: { type: string }
        telefone: { type: string }
        is_admin: { type: boolean }
        data_cadastro: { type: string, format: date-time }
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Autenticação]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, senha]
              properties:
                email: { type: string }
                senha: { type: string }
            example: { email: cliente@lumestore.dev, senha: senha123 }
      responses:
        '200':
          description: Token de sessão (válido por 24 h)
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  usuario: { $ref: '#/components/schemas/Usuario' }
        '401': { description: Credenciais inválidas, content: { application/json: { schema: { $ref: '#/components/schemas/Erro' } } } }
...
  /api/usuarios:
    post:
      tags: [Usuários]
      summary: Criar conta
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome, email, senha]
              properties:
                nome: { type: string }
                email: { type: string }
                senha: { type: string, minLength: 6 }
                endereco: { type: string }
                telefone: { type: string }
            example: { nome: Maria Exemplo, email: maria@lumestore.dev, senha: exemplo123, endereco: 'Rua Exemplo, 45 - Campinas/SP' }
      responses:
        '201': { description: Conta criada, content: { application/json: { schema: { $ref: '#/components/schemas/Usuario' } } } }
        '409': { description: E-mail já cadastrado }
    get:
      tags: [Usuários]
      summary: Listar usuários (admin)
      responses:
        '200': { description: Usuários }
        '403': { description: Não é admin }
...